closes #318 - init commit to refactor all of codebase where possible to use req.user
