+++ graphql/queryResolvers/alerts.ts
@@ -1,0 +1,7 @@
+import db from '../../helpers/dbload'
+
+const { Alert } = db
+
+export const alerts = () => {
+  return Alert.findAll()
+}
