+++ graphql/queryResolvers/lessons.ts
@@ -1,0 +1,13 @@
+import db from '../../helpers/dbload'
+
+const { Lesson } = db
+
+export const lessons = () => {
+  return Lesson.findAll({
+    include: ['challenges'],
+    order: [
+      ['order', 'ASC'],
+      ['challenges', 'order', 'ASC']
+    ]
+  })
+}
